block: Block Device Filtering Mechanism - async snapimage
Can you take a look at the attached patch?  This rewrites the
snapimage I/O so that it does a first order iteration over the
chunks, and then copies from the chunks to the bio_vecs without
an iov_iter.  This should help in making the I/O asynchronous,
but of course can be pretty broken as is and needs separate testing.