remove image_ctx
The algorithm of reading and writing to the snapshot image has been
changed.

In the previous version of the algorithm, data was first read from the
block device to the buffer of the chunk, and only then copied from the
buffer to the bio pages. This led to excessive copying of data from one
memory area to another.

Now data is copied from the buffer only if it already exists in it.
If the buffer is empty, bio clones are created and data is read directly
from the original device or from the difference storage.

If writing to the snapshot image is performed, then the data of the
chunk is read from the original device to the buffer, but only if they
have not yet been read. If the data has already been read from original
block device into the chunks buffer and the chunk is in the cache queue,
then the data is changed in this buffer. If the buffer is empty and the
data has already been stored to the difference storage, then the data is
written directly to the storage.