create file blk-filter.c
All comments from the bdev.c file is completely added to the document
core-api/kernel-api.rst. But the documentation for the functions of
working with filters is placed in a separate section.
So, I think it makes sense to put the code for working with block device
filters in a separate file.