blksnap: attaching and detaching the filter and handling I/O units
The struct tracker contains callback functions for handling a I/O units
of a block device. When a write request is handled, the change block
tracking (CBT) map functions are called and initiates the process of
copying data from the original block device to the change store.
Registering and unregistering the tracker is provided by the functions
blkfilter_register() and blkfilter_unregister().

Signed-off-by: Sergei Shtepa <sergei.shtepa@veeam.com>