blksnap: owner of information about overwritten blocks of the original block device
This is perhaps the key component of the module. It stores information
about the modified blocks of the original device and the location of the
regions where these blocks are stored in the difference storage.
This information allows to restore the state of the block device at the
time of taking the snapshot and represent the snapshot image as a block
device.
When reading from a snapshot, if the block on the original device has
not yet been changed since the snapshot was taken, then the data is read
from the original block device. If the data on the original block device
has been overwritten, then the block is read from the difference
storage. Reads and writes are performed with minimal data storage blocks
(struct chunk).

Signed-off-by: Sergei Shtepa <sergei.shtepa@veeam.com>