add image_io_queue
There is a problem when using the snapshot file. When a request is made
to the file system, requests to the file system are made again at the
filter level of the block device. This leads to deadlocks and increases
the depth of the stack nesting.

Therefore, the processing of the request to the file system by the
block device filter is executed in the system worker.