fix the snapshot corrupt on the system with PAGE_SIZE 64KiB
The problem was in the function chunk_diff_bio(). With a bio of large
size and small chunk sizes, the size of the data for reading/writing to
the difference storage was calculated incorrectly.
The problem was confirmed on a server with a large page size (64KiB).

The function chunk_origin_bio() turned out to be a duplicate of the
function orig_clone_bio(). Was deleted.